Make Rcpp::exception thread-safe
