+++ inst/NEWS.Rd
@@ -7,6 +7,7 @@
   \itemize{
     \item Changes in Rcpp API:
     \itemize{
+      \item \code{Rcpp::exception} is now thread-safe
       \item Safer \code{Rcpp_list*}, \code{Rcpp_lang*} and
       \code{Function.operator()} (Romain in \ghpr{1014}, \ghit{1015}).
       \item A number of \code{#nocov} markers were added (Dirk in
@@ -14,7 +15,7 @@
       \item Finalizer calls clear external pointer first (Kirill Müller and
       Dirk in \ghpr{1038}).
       \item Scalar operations with a rhs matrix no longer change the
-      matrix value (Qiang in \ghpr{1040} fixing (again) \ghit{365}). 
+      matrix value (Qiang in \ghpr{1040} fixing (again) \ghit{365}).
     }
     \item Changes in Rcpp Attributes:
     \itemize{
