+++ inst/include/Rcpp/routines.h
@@ -45,7 +45,6 @@ SEXP          rcpp_set_stack_trace(SEXP);
 std::string   demangle(const std::string& name);
 const char*   short_file_name(const char* );
 int*          get_cache(int n);
-SEXP          stack_trace( const char *file = "", int line = -1);
 SEXP          get_string_elt(SEXP s, R_xlen_t i);
 const char*   char_get_string_elt(SEXP s, R_xlen_t i);
 void          set_string_elt(SEXP s, R_xlen_t i, SEXP v);
@@ -155,12 +154,6 @@ inline attribute_hidden const char* short_file_name(const char* file) {
     typedef const char* (*Fun)(const char*);
     static Fun fun = GET_CALLABLE("short_file_name");
     return fun(file);
-}
-
-inline attribute_hidden SEXP stack_trace( const char *file = "", int line = -1){
-    typedef SEXP (*Fun)(const char*, int);
-    static Fun fun = GET_CALLABLE("stack_trace");
-    return fun(file, line);
 }
 
 inline attribute_hidden SEXP get_string_elt(SEXP s, R_xlen_t i){
